+++ resources/Changelog.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - Increase trader, dog and mutant health and damage on higher difficulties
+- Adjust attack/defense value display to already include experience
 - Camps automatically select better traps if production is too low
 - Show gas and radiation warning in main UI